pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.1.0"
description = "Authentication microservice for AI restaurant platform"
authors = ["Brook Teferi <brookteferi1q2w@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]
package-mode = false   # microservice → no packaging needed

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.1"
uvicorn = { extras = ["standard"], version = "^0.23.2" }
SQLAlchemy = "^2.0.20"
psycopg2-binary = "^2.9.9"
alembic = "^1.12.0"
pydantic = "^2.5.1"
python-dotenv = "^1.0.1"
argon2-cffi = "^21.3.0"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
email-validator = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.0"
ruff = "^0.5.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"